Implemented PHP
I have created an includes folder, and 3 files so far for my header, footer, and sidebar. I have then used PHP tags in my index.php to include those sections of code. I have run this in my terminal with local host and it all seems to be working correctly so far.